Fixed bug with filename case
When the old and the new filenames were equal except for some
letter casings, the rename would fail (atleast in Arch Linux)
because a file with that name (case insensitive) already existed.

Therefore, implemented an extra rename on such files so that before
setting the new name we add a _temp string to the end of the old
file name.

Also added a .gitoverride to the repository